+++ src/dag.rs
@@ -22,6 +22,7 @@ use rust_unixfs::{
     resolve, MaybeResolved,
 };
 use serde::de::DeserializeOwned;
+use serde::Serialize;
 use std::convert::TryFrom;
 use std::error::Error as StdError;
 use std::iter::Peekable;
@@ -201,8 +202,8 @@ impl IpldDag {
     /// Puts an ipld node into the ipfs repo using `dag-cbor` codec and Sha2_256 hash.
     ///
     /// Returns Cid version 1 for the document
-    pub fn put_dag(&self, ipld: Ipld) -> DagPut {
-        self.put().ipld(ipld)
+    pub fn put_dag<S: Serialize>(&self, ipld: S) -> DagPut {
+        self.put().serialize(ipld)
     }
 
     /// Gets an ipld node from the ipfs, fetching the block if necessary.
@@ -226,9 +227,8 @@ impl IpldDag {
         DagGet::new(self.clone())
     }
 
-    pub(crate) async fn get_with_session(
+    pub(crate) async fn _get(
         &self,
-        session: Option<u64>,
         path: IpfsPath,
         providers: &[PeerId],
         local_only: bool,
@@ -255,9 +255,7 @@ impl IpldDag {
         let mut iter = resolved_path.iter().peekable();
 
         let (node, _) = match self
-            .resolve0(
-                session, cid, &mut iter, true, providers, local_only, timeout,
-            )
+            .resolve0(cid, &mut iter, true, providers, local_only, timeout)
             .await
         {
             Ok(t) => t,
@@ -288,13 +286,12 @@ impl IpldDag {
         providers: &[PeerId],
         local_only: bool,
     ) -> Result<(ResolvedNode, SlashedPath), ResolveError> {
-        self.resolve_with_session(None, path, follow_links, providers, local_only, None)
+        self._resolve(path, follow_links, providers, local_only, None)
             .await
     }
 
-    pub(crate) async fn resolve_with_session(
+    pub(crate) async fn _resolve(
         &self,
-        session: Option<u64>,
         path: IpfsPath,
         follow_links: bool,
         providers: &[PeerId],
@@ -322,15 +319,7 @@ impl IpldDag {
         let (node, matched_segments) = {
             let mut iter = resolved_path.iter().peekable();
             match self
-                .resolve0(
-                    session,
-                    cid,
-                    &mut iter,
-                    follow_links,
-                    providers,
-                    local_only,
-                    timeout,
-                )
+                .resolve0(cid, &mut iter, follow_links, providers, local_only, timeout)
                 .await
             {
                 Ok(t) => t,
@@ -353,7 +342,6 @@ impl IpldDag {
     #[allow(clippy::too_many_arguments)]
     async fn resolve0<'a>(
         &self,
-        session: Option<u64>,
         cid: &Cid,
         segments: &mut Peekable<impl Iterator<Item = &'a str>>,
         follow_links: bool,
@@ -371,7 +359,7 @@ impl IpldDag {
         loop {
             let block = match self
                 .repo
-                .get_block_with_session(session, &current, providers, local_only, timeout)
+                ._get_block(&current, providers, local_only, timeout)
                 .await
             {
                 Ok(block) => block,
@@ -442,7 +430,6 @@ impl IpldDag {
 #[must_use = "futures do nothing unless you `.await` or poll them"]
 pub struct DagGet {
     dag_ipld: IpldDag,
-    session: Option<u64>,
     path: Option<IpfsPath>,
     providers: Vec<PeerId>,
     local: bool,
@@ -454,20 +441,12 @@ impl DagGet {
     pub fn new(dag: IpldDag) -> Self {
         Self {
             dag_ipld: dag,
-            session: None,
             path: None,
             providers: vec![],
             local: false,
             timeout: None,
             span: None,
         }
-    }
-
-    /// Bitswap session
-    #[allow(dead_code)]
-    pub(crate) fn session(mut self, session: u64) -> Self {
-        self.session = Some(session);
-        self
     }
 
     /// Path to object
@@ -534,13 +513,7 @@ impl std::future::IntoFuture for DagGet {
         async move {
             let path = self.path.ok_or(ResolveError::PathNotProvided)?;
             self.dag_ipld
-                .get_with_session(
-                    self.session,
-                    path,
-                    &self.providers,
-                    self.local,
-                    self.timeout,
-                )
+                ._get(path, &self.providers, self.local, self.timeout)
                 .await
         }
         .instrument(span)
@@ -598,9 +571,8 @@ impl DagPut {
     }
 
     /// Set a ipld object
-    pub fn ipld(mut self, data: Ipld) -> Self {
-        self.data = Box::new(move || Ok(data));
-        self
+    pub fn ipld(self, data: Ipld) -> Self {
+        self.serialize(data)
     }
 
     /// Set a serde-compatible object