+++ src/unixfs/get.rs
@@ -5,6 +5,7 @@ use std::{
 };
 
 use either::Either;
+use futures::stream::BoxStream;
 use futures::{future::BoxFuture, stream::FusedStream, FutureExt, Stream, StreamExt};
 use libp2p::PeerId;
 #[allow(unused_imports)]
@@ -16,9 +17,9 @@ use tracing::{Instrument, Span};
 use crate::{dag::IpldDag, repo::Repo, Ipfs, IpfsPath};
 
 #[allow(unused_imports)]
-use super::{StatusStreamState, TraversalFailed, UnixfsStatus};
+use super::{TraversalFailed, UnixfsStatus};
 
-#[must_use = "do nothing unless you `.await` or poll the stream"]
+#[must_use = "does nothing unless you `.await` or poll the stream"]
 pub struct UnixfsGet {
     core: Option<Either<Ipfs, Repo>>,
     dest: PathBuf,
@@ -27,7 +28,7 @@ pub struct UnixfsGet {
     providers: Vec<PeerId>,
     local_only: bool,
     timeout: Option<Duration>,
-    stream: StatusStreamState,
+    stream: Option<BoxStream<'static, UnixfsStatus>>,
 }
 
 impl UnixfsGet {
@@ -54,7 +55,7 @@ impl UnixfsGet {
             providers: Vec::new(),
             local_only: false,
             timeout: None,
-            stream: StatusStreamState::None,
+            stream: None,
         }
     }
 
@@ -99,22 +100,10 @@ impl Stream for UnixfsGet {
     ) -> std::task::Poll<Option<Self::Item>> {
         loop {
             match &mut self.stream {
-                StatusStreamState::None => {
-                    let (repo, dag, session) = match self.core.take().expect("ipfs or repo is used")
-                    {
-                        Either::Left(ipfs) => (
-                            ipfs.repo().clone(),
-                            ipfs.dag(),
-                            Some(
-                                crate::BITSWAP_ID.fetch_add(1, std::sync::atomic::Ordering::SeqCst),
-                            ),
-                        ),
-                        Either::Right(repo) => {
-                            let session = repo.is_online().then(|| {
-                                crate::BITSWAP_ID.fetch_add(1, std::sync::atomic::Ordering::SeqCst)
-                            });
-                            (repo.clone(), IpldDag::from(repo.clone()), session)
-                        }
+                None => {
+                    let (repo, dag) = match self.core.take().expect("ipfs or repo is used") {
+                        Either::Left(ipfs) => (ipfs.repo().clone(), ipfs.dag()),
+                        Either::Right(repo) => (repo.clone(), IpldDag::from(repo.clone())),
                     };
 
                     let path = self.path.take().expect("starting point exist");
@@ -125,7 +114,6 @@ impl Stream for UnixfsGet {
 
                     #[cfg(not(target_arch = "wasm32"))]
                     let stream = async_stream::stream! {
-
 
                         let mut cache = None;
                         let mut total_size = None;
@@ -142,7 +130,7 @@ impl Stream for UnixfsGet {
                             };
 
                         let block  = match dag
-                            .resolve_with_session(session, path.clone(), true, &providers, local_only, timeout)
+                            ._resolve(path.clone(), true, &providers, local_only, timeout)
                             .await
                             .map_err(TraversalFailed::Resolving)
                             .and_then(|(resolved, _)| resolved.into_unixfs_block().map_err(TraversalFailed::Path)) {
@@ -160,7 +148,7 @@ impl Stream for UnixfsGet {
 
                         while walker.should_continue() {
                             let (next, _) = walker.pending_links();
-                            let block = match repo.get_block_with_session(session, next, &providers, local_only, timeout).await {
+                            let block = match repo._get_block(next, &providers, local_only, timeout).await {
                                 Ok(block) => block,
                                 Err(e) => {
                                     yield UnixfsStatus::FailedStatus { written, total_size, error: Some(e) };
@@ -221,7 +209,6 @@ impl Stream for UnixfsGet {
                     let stream = async_stream::stream! {
                         _ = repo;
                         _ = dag;
-                        _ = session;
                         _ = path;
                         _ = providers;
                         _ = local_only;
@@ -230,29 +217,24 @@ impl Stream for UnixfsGet {
                         yield UnixfsStatus::FailedStatus { written: 0, total_size: None, error: Some(anyhow::anyhow!("unimplemented")) };
                     };
 
-                    self.stream = StatusStreamState::Pending {
-                        stream: stream.boxed(),
-                    };
+                    self.stream = Some(stream.boxed());
                 }
-                StatusStreamState::Pending { stream } => {
-                    match futures::ready!(stream.poll_next_unpin(cx)) {
-                        Some(item) => {
-                            if matches!(
-                                item,
-                                UnixfsStatus::FailedStatus { .. }
-                                    | UnixfsStatus::CompletedStatus { .. }
-                            ) {
-                                self.stream = StatusStreamState::Done;
-                            }
-                            return Poll::Ready(Some(item));
+                Some(stream) => match futures::ready!(stream.poll_next_unpin(cx)) {
+                    Some(item) => {
+                        if matches!(
+                            item,
+                            UnixfsStatus::FailedStatus { .. }
+                                | UnixfsStatus::CompletedStatus { .. }
+                        ) {
+                            self.stream.take();
                         }
-                        None => {
-                            self.stream = StatusStreamState::Done;
-                            return Poll::Ready(None);
-                        }
+                        return Poll::Ready(Some(item));
                     }
-                }
-                StatusStreamState::Done => return Poll::Ready(None),
+                    None => {
+                        self.stream.take();
+                        return Poll::Ready(None);
+                    }
+                },
             }
         }
     }
@@ -286,6 +268,6 @@ impl std::future::IntoFuture for UnixfsGet {
 
 impl FusedStream for UnixfsGet {
     fn is_terminated(&self) -> bool {
-        matches!(self.stream, StatusStreamState::Done) && self.core.is_none()
+        self.stream.is_none() && self.core.is_none()
     }
 }