+++ src/unixfs/ls.rs
@@ -21,7 +21,7 @@ pub enum Entry {
     File { cid: Cid, file: String, size: usize },
 }
 
-#[must_use = "do nothing unless you `.await` or poll the stream"]
+#[must_use = "does nothing unless you `.await` or poll the stream"]
 pub struct UnixfsLs {
     core: Option<Either<Ipfs, Repo>>,
     span: Span,
@@ -107,20 +107,9 @@ impl Stream for UnixfsLs {
                         return Poll::Ready(None);
                     };
 
-                    let (repo, dag, session) = match core {
-                        Either::Left(ipfs) => (
-                            ipfs.repo().clone(),
-                            ipfs.dag(),
-                            Some(
-                                crate::BITSWAP_ID.fetch_add(1, std::sync::atomic::Ordering::SeqCst),
-                            ),
-                        ),
-                        Either::Right(repo) => {
-                            let session = repo.is_online().then(|| {
-                                crate::BITSWAP_ID.fetch_add(1, std::sync::atomic::Ordering::SeqCst)
-                            });
-                            (repo.clone(), IpldDag::from(repo.clone()), session)
-                        }
+                    let (repo, dag) = match core {
+                        Either::Left(ipfs) => (ipfs.repo().clone(), ipfs.dag()),
+                        Either::Right(repo) => (repo.clone(), IpldDag::from(repo.clone())),
                     };
 
                     let path = self.path.take().expect("path exist");
@@ -133,7 +122,7 @@ impl Stream for UnixfsLs {
                     let stream = async_stream::stream! {
 
                         let resolved = match dag
-                            .resolve_with_session(session, path, true, &providers, local_only, timeout)
+                            ._resolve(path, true, &providers, local_only, timeout)
                             .await {
                                 Ok((resolved, _)) => resolved,
                                 Err(e) => {
@@ -158,7 +147,7 @@ impl Stream for UnixfsLs {
                         let mut root_directory = String::new();
                         while walker.should_continue() {
                             let (next, _) = walker.pending_links();
-                            let block = match repo.get_block_with_session(session, next, &providers, local_only, timeout).await {
+                            let block = match repo._get_block(next, &providers, local_only, timeout).await {
                                 Ok(block) => block,
                                 Err(error) => {
                                     yield Entry::Error { error };